[revscriptsys] adding ItemType properties (#8)
Adding ItemType properties from `items.otb` in favour of removing both `items.xml` and `items.otb` and instead make individual `.lua` files for each itemid (possibility to fetch either items as serverid or clientid)
This PR wont include the remove of `items.xml` and `items.otb`. 
I want to do that in a separate PR for a better overview.