Items.xml & Items.otb removal (#9)
We are dropping support for `Items.xml` and `Items.otb` in favor of creating items through lua scripts directly.
Now we can create/manipulate ItemTypes on the fly without the need to keep `Items.otb` and `Items.xml` in sync. 
Properties are now saved directly in the lua file.
For future release I'll add the option to chose in `config.lua` if we want to load items by `server id` or by `client id` this way we can leave us options open which type we prefer.